feat!: Remove Google Bastion module, provider >= 5
In prior releases this module was wrapper for Google's bastion module,
just with a more limited set of options. This has caused problems when
composing environments when the upstream Google module depended on
values only known/computed at apply. This change removes the Google
module for a simpler, but more specialised form of bastion - single VM,
IAP only.

This module has changed the minimal supported versions of the following:
 - Google Terraform provider >= 5.0
 - Terraform >= 1.2

In addition some of the input variables have changed. Please review
docs before updating.

Closes #73.